server/requirements.txt
fastapi
cloudinary
sqlalchemy
dotenv
bcrypt
pydantic-core
PyJWT